Refactor agent implementation: consolidate agent classes into a unified structure, enhance system message loading, and update configuration handling in Bicep and Python files.
