+++ src/agents/__init__.py
@@ -1,0 +1,8 @@
+# This file makes the agents directory a Python package
+
+from .agent import Agent
+from .agent_registry import agent_registry
+from .multi_agent import MultiAgent
+from .single_agent import SingleAgent
+
+__all__ = ["agent_registry", "Agent", "SingleAgent", "MultiAgent"]
