+++ src/agents/agent_registry.py
@@ -24,167 +24,19 @@ agent_registry:
 """
 
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Type, Union
+from typing import Dict, Optional
 
 import yaml
 from streamlit.logger import get_logger
 
-PAGES_FILE = Path(__file__).parent / "config/pages.yaml"
+from .agent import Agent
+from .multi_agent import MultiAgent
+from .single_agent import SingleAgent
+
+PAGES_FILE = Path(__file__).parent.parent / "config/pages.yaml"
+
 
 logger = get_logger(__name__)
-
-
-class Agent:
-    """
-    Base class for agent implementations.
-
-    Attributes:
-    -----------
-    agent_key : str
-        Unique identifier for the agent.
-    model : str
-        The model to use for this agent.
-    temperature : float
-        The temperature setting for response generation.
-    """
-
-    def __init__(
-        self, agent_key: str, model: str = "gpt-4o", temperature: float = 0.7
-    ) -> None:
-        """
-        Initialize an Agent with configurable settings.
-
-        Parameters:
-        -----------
-        agent_key : str
-            Unique identifier for the agent.
-        model : str, optional
-            The model to use for this agent. Defaults to "gpt-4o".
-        temperature : float, optional
-            The temperature setting for response generation. Defaults to 0.7.
-        """
-        self.agent_key = agent_key
-        self.model = model
-        self.temperature = temperature
-
-    def get_system_messages(self) -> List[Dict[str, str]]:
-        """
-        Get the system messages for this agent.
-
-        Returns:
-        --------
-        List[Dict[str, str]]
-            A list of system messages for the agent.
-        """
-        raise NotImplementedError("Subclasses must implement this method")
-
-
-class SingleAgent(Agent):
-    """
-    Implementation of a single agent with one persona.
-
-    Attributes:
-    -----------
-    persona : str
-        Path to the persona prompt file.
-    documents : Optional[Union[str, List[str]]]
-        Path(s) to document context file(s).
-    """
-
-    def __init__(
-        self,
-        agent_key: str,
-        persona: str,
-        model: str = "gpt-4o",
-        documents: Optional[Union[str, List[str]]] = None,
-        temperature: float = 0.7,
-    ) -> None:
-        """
-        Initialize a SingleAgent with a specific persona.
-
-        Parameters:
-        -----------
-        agent_key : str
-            Unique identifier for the agent.
-        persona : str
-            Path to the persona prompt file.
-        model : str, optional
-            The model to use for this agent. Defaults to "gpt-4o".
-        documents : Optional[Union[str, List[str]]], optional
-            Path(s) to document context file(s). Defaults to None.
-        temperature : float, optional
-            The temperature setting for response generation. Defaults to 0.7.
-        """
-        super().__init__(agent_key, model, temperature)
-        self.persona = persona
-        self.documents = documents
-
-    def get_system_messages(self) -> List[Dict[str, str]]:
-        """
-        Get the system messages for this agent based on persona and documents.
-
-        Returns:
-        --------
-        List[Dict[str, str]]
-            A list of system messages for the agent.
-        """
-        # This will be implemented using the load_prompt_files function
-        # For now, we're just defining the interface
-        pass
-
-
-class MultiAgent(Agent):
-    """
-    Implementation of multiple agents with multiple personas.
-
-    Attributes:
-    -----------
-    personas : List[str]
-        List of paths to the persona prompt files.
-    documents : Optional[List[str]]
-        List of paths to document context files.
-    """
-
-    def __init__(
-        self,
-        agent_key: str,
-        personas: List[str],
-        model: str = "gpt-4o",
-        documents: Optional[List[str]] = None,
-        temperature: float = 0.7,
-    ) -> None:
-        """
-        Initialize a MultiAgent with multiple personas.
-
-        Parameters:
-        -----------
-        agent_key : str
-            Unique identifier for the agent.
-        personas : List[str]
-            List of paths to the persona prompt files.
-        model : str, optional
-            The model to use for this agent. Defaults to "gpt-4o".
-        documents : Optional[List[str]], optional
-            List of paths to document context files. Defaults to None.
-        temperature : float, optional
-            The temperature setting for response generation. Defaults to 0.7.
-        """
-        super().__init__(agent_key, model, temperature)
-        self.personas = personas
-        self.documents = documents
-
-    def get_system_messages(self) -> List[Dict[str, str]]:
-        """
-        Get the system messages for this agent based on personas and documents.
-
-        Returns:
-        --------
-        List[Dict[str, str]]
-            A list of system messages for the agent.
-        """
-        # This will be implemented using the load_prompt_files function
-        # For now, we're just defining the interface
-        pass
 
 
 class AgentRegistry:
@@ -255,8 +107,6 @@ class AgentRegistry:
         ValueError:
             If the agent configuration is invalid or missing required fields.
         """
-        # Import here to avoid circular imports
-        from agent_impl import MultiAgentImpl, SingleAgentImpl
 
         agent_config = self.get(agent_key)
         if not agent_config:
@@ -281,7 +131,7 @@ class AgentRegistry:
                     agent_key,
                     model,
                 )
-                return SingleAgentImpl(
+                return SingleAgent(
                     agent_key=agent_key,
                     persona=agent_config["persona"],
                     model=model,
@@ -293,7 +143,7 @@ class AgentRegistry:
                 logger.info(
                     "Creating MultiAgent for key '%s' with model '%s'", agent_key, model
                 )
-                return MultiAgentImpl(
+                return MultiAgent(
                     agent_key=agent_key,
                     personas=agent_config["personas"],
                     model=model,
